+++ src/main/java/com/solvd/interfaces/ITax.java
@@ -3,5 +3,5 @@ package com.solvd.interfaces;
 public interface ITax {
 
     // Add country's taxes
-    float tax(float price);
+    Double tax(Double price);
 }
